MSCR-299 Refactor content adding forms, midway save
