+++ mscr-ui/src/common/interfaces/format.interface.ts
@@ -1,0 +1,9 @@
+export type Format =
+  | 'JSONSCHEMA'
+  | 'CSV'
+  | 'SKOSRDF'
+  | 'PDF'
+  | 'XSD'
+  | 'XML'
+  | 'XLST'
+  | 'MSCR';
