+++ mscr-ui/src/common/interfaces/schema.interface.ts
@@ -1,3 +1,7 @@
+import { State } from '@app/common/interfaces/state.interface';
+import { Format } from '@app/common/interfaces/format.interface';
+import { LanguageBlockType } from 'yti-common-ui/components/form/language-selector';
+
 export interface Schema {
   namespace?: string;
   pid?: string;
@@ -55,15 +59,10 @@ export interface Organization {
 // ToDo: Proper typing
 export interface SchemaFormType {
   namespace?: string;
-  contact?: boolean;
-  serviceCategories?: any;
   pid?: string;
-  format?: string;
-  label?: { [key: string]: string };
-  languages: any;
-  organizations: any;
+  format: Format;
+  languages: (LanguageBlockType & { selected: boolean })[];
+  organizations: Organization[];
   filedata?: any;
-  description?: any;
-  status?: any; //Status will be set to schema.state
-  uri?: any;
+  state: State;
 }
