+++ mscr-ui/src/common/interfaces/state.interface.ts
@@ -1,0 +1,6 @@
+export type State =
+  | 'DRAFT'
+  | 'PUBLISHED'
+  | 'INVALID'
+  | 'DEPRECATED'
+  | 'REMOVED';
