--- mscr-ui/src/common/interfaces/status.interface.ts
@@ -1,1 +1,0 @@
-export type Status = 'DRAFT' | 'PUBLISHED' | 'DEPRECATED';
