+++ mscr-ui/src/modules/form/validate-schema-form.tsx
@@ -1,6 +1,7 @@
 import { SchemaFormType } from '@app/common/interfaces/schema.interface';
 
 // Not yet modified according to mscr validation errors
+// Right now checks if selected languages have missing titles, or there is no file data
 export interface FormErrors {
   languageAmount: boolean;
   titleAmount: string[];
@@ -9,7 +10,7 @@ export interface FormErrors {
   fileData: boolean;
 }
 
-export function validateForm(
+export function validateSchemaForm(
   data: SchemaFormType,
   fileData: File | null | undefined
 ) {
@@ -23,7 +24,7 @@ export function validateForm(
   };
 
   const selectedLanguages = data.languages.filter(
-    (lang: { selected: any }) => lang.selected
+    (lang: { selected: boolean }) => lang.selected
   );
 
   // At least one language should be selected
@@ -32,18 +33,14 @@ export function validateForm(
   }
 
   // All selected languages should have a title
+  const titleless = selectedLanguages.filter(
+    (lang: { title: string | undefined }) =>
+      !lang.title || lang.title === '' || lang.title.length < 1
+  );
   if (
-    selectedLanguages.filter(
-      (lang: { title: string | any[] }) =>
-        !lang.title || lang.title === '' || lang.title.length < 1
-    ).length > 0
+    titleless.length > 0
   ) {
-    const langsWithError = selectedLanguages
-      .filter(
-        (lang: { title: string | any[] }) =>
-          !lang.title || lang.title === '' || lang.title.length < 1
-      )
-      .map((lang: { uniqueItemId: any }) => lang.uniqueItemId);
+    const langsWithError = titleless.map((lang: { uniqueItemId: string }) => lang.uniqueItemId);
 
     errors.titleAmount = langsWithError ?? [];
   }
