+++ mscr-ui/src/modules/workspace/personal-home/index.tsx
@@ -9,12 +9,11 @@ import {
 } from 'yti-common-ui/components/title/title.styles';
 import Separator from 'yti-common-ui/components/separator';
 import { ButtonBlock } from '@app/modules/front-page/front-page.styles';
-import SchemaFormModal from '@app/modules/schema-form/schema-form-modal';
+import SchemaFormModal from '@app/modules/form/schema-form-modal';
 import { useBreakpoints } from 'yti-common-ui/components/media-query';
 import { useGetOrganizationsQuery } from '@app/common/components/organizations/organizations.slice';
-import CrosswalkFormModal from '@app/modules/crosswalk-form/crosswalk-form-modal';
+import CrosswalkFormModal from '@app/modules/form/crosswalk-form-modal';
 import CrosswalkSelectionModal from '@app/modules/create-crosswalk/crosswalk-selection-modal';
-import RegisterSchemaModal from '@app/modules/form/register-schema-modal';
 
 export default function PersonalWorkspace({
   contentType,
@@ -54,7 +53,6 @@ export default function PersonalWorkspace({
           {contentType == 'SCHEMA' ? (
             <>
               <SchemaFormModal refetch={refetchInfo}></SchemaFormModal>
-              <RegisterSchemaModal />
             </>
           ) : (
             <>
